+++ xarray/datatree_/docs/source/whats-new.rst
@@ -40,6 +40,9 @@ Bug fixes
 Documentation
 ~~~~~~~~~~~~~
 
+- Added ``Reading and Writing Files`` page. (:pull:`158`)
+  By `Tom Nicholas <https://github.com/TomNicholas>`_.
+
 Internal Changes
 ~~~~~~~~~~~~~~~~